Mathch_wins.py
Alice and Bob are playing a bastketball game. Each ball scored earns 2 points, but if it is scored from strictly more than 6.75 units away, it earns 3 points. Help them to find out who won the game!
The basket is located at coordinate (0,0).
Input
Line 1 : An integer N for the number of coordinates to consider.
N next lines : A player player (A for Alice and B for Bob) followed by two float x and y for the coordinates of where he/she scored a basket separated by a space.
Output
The winner of the match, or DRAW
Constraints
0 ≤ N ≤ 10
0 ≤ x ≤ 100
-100 ≤ y ≤ 100
Example
Input
1
A 1 1
Output
ALICE