Sweet_box.java
In mathematics exam N students passed and some others failed !
The teacher John decided to give sweets to students .He will distribute the sweets to students who passed the exam in the order of 1 sweet for first student 3 sweets for second student 5 sweets for third student and so on ...

the sweets are in packets. Each packet contains 10 sweets. how many packets are needed for giving sweets for students. He said that next time the failed students will study because of this !
Input
Line 1 : an integer N the number of students who passed the exam
Output
Line 1 : P the number of packets needed
Constraints
10 ≤ N ≤ 100
10 ≤ P ≤ 1000
Example
Input
10
Output
10